2007-06-19  Hans Boehm <Hans.Boehm@hp.com>
	* alloc.c (GC_adj_bytes_allocd): Avoid (long) casts, fix comment.
	(GC_print_heap_sects): Use size_t instead of unsigned long.
	* thread_local_alloc.c (GC_lookup_thread): Define in the correct
	context.
	* win32_threads.c, include/gc_config_macros.h: The last of Romano
	Paolo Tenca's patch.  Move stdint.h include to gc_config_macros.h.
	* include/gc_inline.h: Avoid gc_priv.h dependencies.
	* tests/test.c (check_heap_stats): Replace unsigned long with size_t.

+++ tests/test.c
@@ -1169,7 +1169,7 @@ void run_one_test()
 
 void check_heap_stats()
 {
-    unsigned long max_heap_sz;
+    size_t max_heap_sz;
     int i;
     int still_live;
     int late_finalize_count = 0;
